requirements.txt
aiohttp==3.7.4.post0
alembic==1.5.8
anyio==2.2.0
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
Babel==2.9.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
bleach==3.3.0
blueprint==3.4.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==5.0.3
defusedxml==0.7.1
dnspython==2.0.0
email-validator==1.1.1
entrypoints==0.3
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-CKEditor==0.4.6
Flask-Login==0.4.1
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
geoip2==4.1.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.10.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.5.1
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==1.1.1
maxminddb==2.0.3
mccabe==0.6.1
mistune==0.8.4
multidict==5.1.0
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.19.2
opencv-python==4.4.0.44
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.10.0
prompt-toolkit==3.0.18
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.8.1
pylint==2.5.3
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
requests==2.24.0
scapy==2.4.4
schedule==0.6.0
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.2.1
SQLAlchemy==1.3.11
terminado==0.9.4
testpath==0.4.4
toml==0.10.1
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.12.1
WTForms==2.1
yarl==1.6.3
zipp==3.4.1
